+++ code/scripting/doc_html.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "scripting_doc.h"
+
+namespace scripting {
+
+void output_html_doc(const ScriptingDocumentation& doc, const SCP_string& filename);
+
+}
