+++ code/scripting/doc_json.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "scripting_doc.h"
+
+namespace scripting {
+
+void output_json_doc(const ScriptingDocumentation& doc, const SCP_string& filename);
+
+}
